Sources and File Names (ShadowChemosh)
I changed around the .1st file sources to better support all the
different Path of War books. In addition made helper Thing changes so
that we use the Parent source. This way if ANY child source is active
the base Path of War Things display or activate. Found some archetypes
that where not correct source marked and fixed those.

I changed the .user file name to better match into the layout for the
current files.