Unique ID's
I changed the Unique ID's of pretty much everything in the .user file.
We HAVE to be VERY careful about unique ID's as they have to be unique
accross ALL of Hero Lab.  This means unique vs every single add on now
and the future. Plus unquie vs everythign ever to be added by Paizo.

This is why we went to a number system of PWx where X is the next class
number (ie 7 for Medic). This gives is VERY little chance of running
into a duplicate Unique ID and causing HL to go into Safe Mode.