WIP - Add Divergent Paths: Medic (#84)
* Add Divergent Paths: Medic source

* add a first pass at the medic

* Fix maneuvers not showing up

* Fix Witch Doctor Skills

* Sources and File Names (ShadowChemosh)

I changed around the .1st file sources to better support all the
different Path of War books. In addition made helper Thing changes so
that we use the Parent source. This way if ANY child source is active
the base Path of War Things display or activate. Found some archetypes
that where not correct source marked and fixed those.

I changed the .user file name to better match into the layout for the
current files.

* Archetype Headers(ShadowChemosh)

Centralized the archetype header Things into a single place to make it
easier to work with.

* Unique ID's

I changed the Unique ID's of pretty much everything in the .user file.
We HAVE to be VERY careful about unique ID's as they have to be unique
accross ALL of Hero Lab.  This means unique vs every single add on now
and the future. Plus unquie vs everythign ever to be added by Paizo.

This is why we went to a number system of PWx where X is the next class
number (ie 7 for Medic). This gives is VERY little chance of running
into a duplicate Unique ID and causing HL to go into Safe Mode.

* Product ID Tags

-Added Product ID to all Things.
-Added source book info to class and archetypes.
-Removed Source from the Class Specials. We don't source mark Class
Specials or Racial Specials because they get bootstrapped to something
that is source marked. This allows us to re-use these abiliies accross
the Pack without having to care whitch Source is turned on.
-Removed all pickreq as its not user friendly. Does not always work
right.  Please always use Expr-Reqs or Pre-Reqs scripts.

Done for today. Leaving off at ar3PAmbLan.