ShadowChemosh Adjustment (ShadowChemosh)
This adjustment has been enhanced with a dropdown to reset with nights
rest or no reset.  It defaults to no reset which is what it does
currently.

In addition a usage dropdown selection on the adjustmnet lets you set
the tracker to be [B]/day[/B] or [B]/week[/B] or whatever.

Finally removed the dropdown to select the tracker.  Instead the tracker
on the In-Play tab gets its data directly from the attached adjustment.