Path of War (ShadowChemosh)
-Fixed #1 LONG TIME outstanding issues with selecting manuevers.  LW
added a new feature that lets me finally check the IL of a class to
validate if a manuever/stance is valid to be added to a character.  This
means that when level 1 PoW class and selecting manuevers it will not
display levels 2+ as being valid to add to your character.  Thanks LW
for this new feature!